Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/xhr/access-control-preflight-request-must-not-contain-cookie-expected.txt


FAIL Preflight request must not contain any cookie header assert_equals: expected "COOKIE" but got ""
